+++ rpc/utils.go
@@ -62,7 +62,7 @@ func EstimateTip(
 	tip U64,
 	err error,
 ) {
-	rawLatestBlock, err := provider.BlockWithTxs(ctx, WithBlockTag(BlockTagLatest))
+	rawLatestBlock, err := provider.BlockWithTxs(ctx, WithBlockTag(BlockTagLatest), nil)
 	if err != nil {
 		return tip, fmt.Errorf("failed to get latest block: %w", err)
 	}